Updated initialization in core and util. Updated the join_radar to use new datetime utilities. Moved to_vpt, is_vpt to radar_utils. Updated radardisplay to correct bugs.
