+++ pyart/util/__init__.py
@@ -42,5 +42,6 @@ from .circular_stats import interval_mean, interval_std
 from .circular_stats import mean_of_two_angles, mean_of_two_angles_deg
 from .xsect import cross_section_ppi
 from .hildebrand_sekhon import estimate_noise_hs74
+from .radar_utils import is_vpt, to_vpt
 
 __all__ = [s for s in dir() if not s.startswith('_')]
